requirements/common.txt
Django==2.1.4
django-crispy-forms==1.7.2
django-environ==0.4.5
django-filter==2.0.0
django-tables2==2.0.3
django-treebeard==4.3
numpy==1.15.4
pandas==0.23.4
pydicom==1.2.1
python-dateutil==2.7.5
pytz==2018.7
xlrd==1.2.0
psycopg2-binary==2.7.6.1
Pillow==5.4.0